Fusion User and Member, remove registration
