--- src/Controller/MemberController.php
@@ -1,117 +1,0 @@
-<?php
-
-/*
- * Copyright 2020 Mathieu Piot
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace App\Controller;
-
-use App\Entity\Member;
-use App\Form\MemberType;
-use App\Repository\InvitationRepository;
-use App\Repository\MemberRepository;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-use Symfony\Component\Routing\Annotation\Route;
-
-/**
- * @Route("/member")
- * @Security("is_granted('ROLE_ADMIN')")
- */
-class MemberController extends AbstractController
-{
-    /**
-     * @Route("/", name="member_index", methods={"GET"})
-     */
-    public function index(Request $request, MemberRepository $memberRepository): Response
-    {
-        return $this->render('member/index.html.twig', [
-            'members' => $memberRepository->findPaginated(
-                $request->query->getAlnum('q'),
-                $request->query->getInt('page', 1)
-            ),
-        ]);
-    }
-
-    /**
-     * @Route("/new", name="member_new", methods={"GET","POST"})
-     */
-    public function new(Request $request): Response
-    {
-        $member = new Member();
-        $form = $this->createForm(MemberType::class, $member);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->persist($member);
-            $entityManager->flush();
-
-            return $this->redirectToRoute('member_index');
-        }
-
-        return $this->render('member/new.html.twig', [
-            'member' => $member,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @Route("/{id}", name="member_show", methods={"GET"})
-     */
-    public function show(Member $member, InvitationRepository $invitationRepository): Response
-    {
-        return $this->render('member/show.html.twig', [
-            'member' => $member,
-            'invitation' => $invitationRepository->findOneBy(['member' => $member]),
-        ]);
-    }
-
-    /**
-     * @Route("/{id}/edit", name="member_edit", methods={"GET","POST"})
-     */
-    public function edit(Request $request, Member $member): Response
-    {
-        $form = $this->createForm(MemberType::class, $member);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('member_index');
-        }
-
-        return $this->render('member/edit.html.twig', [
-            'member' => $member,
-            'form' => $form->createView(),
-        ]);
-    }
-
-    /**
-     * @Route("/{id}", name="member_delete", methods={"DELETE"})
-     */
-    public function delete(Request $request, Member $member): Response
-    {
-        if ($this->isCsrfTokenValid('delete'.$member->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($member);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('member_index');
-    }
-}
