--- src/Entity/Invitation.php
@@ -1,67 +1,0 @@
-<?php
-
-/*
- * Copyright 2020 Mathieu Piot
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace App\Entity;
-
-use App\Utils\CodeGenerator;
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Entity(repositoryClass="App\Repository\InvitationRepository")
- */
-class Invitation
-{
-    /**
-     * @ORM\Id()
-     * @ORM\GeneratedValue()
-     * @ORM\Column(type="integer")
-     */
-    private $id;
-
-    /**
-     * @ORM\ManyToOne(targetEntity="App\Entity\Member")
-     * @ORM\JoinColumn(nullable=false)
-     */
-    private $member;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    private $code;
-
-    public function __construct(Member $member)
-    {
-        $this->member = $member;
-        $this->code = CodeGenerator::generateCode();
-    }
-
-    public function getId(): ?int
-    {
-        return $this->id;
-    }
-
-    public function getMember(): ?Member
-    {
-        return $this->member;
-    }
-
-    public function getCode(): ?string
-    {
-        return $this->code;
-    }
-}
