+++ src/Entity/User.php
@@ -18,20 +18,34 @@
 
 namespace App\Entity;
 
-use App\Validator\UniqueUserMember;
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 use Symfony\Component\Security\Core\User\UserInterface;
+use Symfony\Component\String\Slugger\AsciiSlugger;
 use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Validator\Context\ExecutionContextInterface;
+use function Symfony\Component\String\u;
 
 /**
  * @ORM\Entity(repositoryClass="App\Repository\UserRepository")
  * @ORM\Table(name="app_user")
- * @UniqueEntity(fields={"email"}, message="Un compte existe déjà avec cet email.")
- * @UniqueUserMember()
+ * @UniqueEntity(fields={"firstName", "lastName"}, message="Un compte existe déjà avec ce nom et prénom.")
+ * @UniqueEntity(fields={"licenseNumber"}, message="Un compte existe déjà avec ce numéro de licence.")
+ * @ORM\HasLifecycleCallbacks()
  */
 class User implements UserInterface
 {
+    const NUM_ITEMS = 20;
+
+    const ROWER_CATEGORY_A = 1;
+    const ROWER_CATEGORY_B = 2;
+    const ROWER_CATEGORY_C = 3;
+
+    const LICENSE_TYPE_ANNUAL = 'A';
+    const LICENSE_TYPE_INDOOR = 'I';
+
     /**
      * @ORM\Id()
      * @ORM\GeneratedValue()
@@ -41,29 +55,160 @@ class User implements UserInterface
 
     /**
      * @ORM\Column(type="string", length=180, unique=true)
+     */
+    private $username;
+
+    /**
+     * @ORM\Column(type="string", length=180)
+     * @Assert\NotBlank()
      * @Assert\Email()
      */
     private $email;
 
     /**
-     * @ORM\Column(type="json")
+     * @ORM\Column(type="json", nullable=true)
      */
     private $roles = [];
 
     /**
      * @var string The hashed password
-     * @ORM\Column(type="string")
+     * @ORM\Column(type="string", nullable=true)
      */
     private $password;
 
     /**
-     * @ORM\OneToOne(targetEntity="App\Entity\Member", mappedBy="user")
+     * @ORM\Column(type="string", length=255)
+     * @Assert\NotBlank()
      */
-    private $member;
+    private $gender;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     * @Assert\NotBlank()
+     */
+    private $firstName;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     * @Assert\NotBlank()
+     */
+    private $lastName;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true, unique=true)
+     */
+    private $licenseNumber;
+
+    /**
+     * @ORM\Column(type="date", nullable=true)
+     */
+    private $licenseEndAt;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     * @Assert\NotBlank()
+     */
+    private $licenseType;
+
+    /**
+     * @ORM\Column(type="integer")
+     * @Assert\NotBlank()
+     */
+    private $rowerCategory;
+
+    /**
+     * @ORM\Column(type="date")
+     * @Assert\NotBlank()
+     */
+    private $birthday;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     */
+    private $legalRepresentative;
+
+    /**
+     * @ORM\Column(type="date")
+     * @Assert\NotBlank()
+     */
+    private $subscriptionDate;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Assert\NotBlank()
+     */
+    private $laneNumber;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Assert\NotBlank()
+     */
+    private $laneType;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Assert\NotBlank()
+     */
+    private $laneName;
+
+    /**
+     * @ORM\Column(type="integer", nullable=true)
+     * @Assert\NotBlank()
+     */
+    private $postalCode;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Assert\NotBlank()
+     */
+    private $city;
+
+    /**
+     * @ORM\Column(type="string", length=255, nullable=true)
+     * @Assert\NotBlank()
+     */
+    private $phoneNumber;
+
+    /**
+     * @ORM\ManyToMany(targetEntity="App\Entity\LogbookEntry", mappedBy="crewMembers")
+     */
+    private $logbookEntries;
+
+    /**
+     * @ORM\OneToMany(targetEntity="App\Entity\MedicalCertificate", mappedBy="user", cascade={"persist", "remove"}, orphanRemoval=true)
+     * @ORM\OrderBy({"date": "desc"})
+     * @Assert\Valid()
+     */
+    private $medicalCertificates;
+
+    public function __construct()
+    {
+        $this->licenseEndAt = new \DateTime('last day of october next year');
+        $this->licenseType = self::LICENSE_TYPE_ANNUAL;
+        $this->rowerCategory = self::ROWER_CATEGORY_C;
+        $this->logbookEntries = new ArrayCollection();
+        $this->subscriptionDate = new \DateTimeImmutable();
+        $this->medicalCertificates = new ArrayCollection();
+    }
+
+    public function __toString()
+    {
+        return $this->getFullName();
+    }
 
     public function getId(): ?int
     {
         return $this->id;
+    }
+
+    /**
+     * A visual identifier that represents this user.
+     *
+     * @see UserInterface
+     */
+    public function getUsername(): ?string
+    {
+        return $this->username;
     }
 
     public function getEmail(): ?string
@@ -73,19 +218,9 @@ class User implements UserInterface
 
     public function setEmail(string $email): self
     {
-        $this->email = mb_strtolower($email);
+        $this->email = u($email)->lower();
 
         return $this;
-    }
-
-    /**
-     * A visual identifier that represents this user.
-     *
-     * @see UserInterface
-     */
-    public function getUsername(): string
-    {
-        return (string) $this->email;
     }
 
     /**
@@ -139,21 +274,353 @@ class User implements UserInterface
         // $this->plainPassword = null;
     }
 
-    public function getMember(): ?Member
+
+    public function getGender(): ?string
     {
-        return $this->member;
+        return $this->gender;
     }
 
-    public function setMember(?Member $member): self
+    public function getTextGender(): ?string
     {
-        $this->member = $member;
+        if ('f' === $this->gender) {
+            return 'Femme';
+        }
 
-        // set (or unset) the owning side of the relation if necessary
-        $newUser = null === $member ? null : $this;
-        if ($member->getUser() !== $newUser) {
-            $member->setUser($newUser);
+        return 'Homme';
+    }
+
+    public function setGender(string $gender): self
+    {
+        $this->gender = $gender;
+
+        return $this;
+    }
+
+    public function getFirstName(): ?string
+    {
+        return $this->firstName;
+    }
+
+    public function setFirstName(string $firstName): self
+    {
+        $this->firstName = $firstName;
+
+        return $this;
+    }
+
+    public function getLastName(): ?string
+    {
+        return $this->lastName;
+    }
+
+    public function setLastName(string $lastName): self
+    {
+        $this->lastName = $lastName;
+
+        return $this;
+    }
+
+    public function getFullName(): string
+    {
+        return $this->getFirstName().' '.$this->getLastName();
+    }
+
+    public function getLicenseNumber(): ?string
+    {
+        return $this->licenseNumber;
+    }
+
+    public function setLicenseNumber(?string $licenseNumber): self
+    {
+        $this->licenseNumber = $licenseNumber;
+
+        return $this;
+    }
+
+    public function getLicenseEndAt(): ?\DateTimeInterface
+    {
+        return $this->licenseEndAt;
+    }
+
+    public function setLicenseEndAt(?\DateTimeInterface $licenseEndAt): self
+    {
+        $this->licenseEndAt = $licenseEndAt;
+
+        return $this;
+    }
+
+    public function isLicenseValid(): bool
+    {
+        $now = (new \DateTime())->setTime(0, 0, 0, 0);
+
+        if ($this->licenseEndAt >= $now) {
+            return true;
+        }
+
+        return false;
+    }
+
+    public function getLicenseType(): ?string
+    {
+        return $this->licenseType;
+    }
+
+    public function setLicenseType(?string $licenseType): self
+    {
+        $this->licenseType = $licenseType;
+
+        return $this;
+    }
+
+    public function getTextLicenseType(): ?string
+    {
+        return array_flip(self::getAvailableLicenseTypes())[$this->licenseType];
+    }
+
+    public function getRowerCategory(): ?int
+    {
+        return $this->rowerCategory;
+    }
+
+    public function setRowerCategory(int $rowerCategory): self
+    {
+        $this->rowerCategory = $rowerCategory;
+
+        return $this;
+    }
+
+    public function getTextRowerCategory(): string
+    {
+        return array_flip(self::getAvailableRowerCategories())[$this->rowerCategory];
+    }
+
+    public function getBirthday(): ?\DateTimeInterface
+    {
+        return $this->birthday;
+    }
+
+    public function setBirthday(?\DateTimeInterface $birthday): self
+    {
+        $this->birthday = $birthday;
+
+        return $this;
+    }
+
+    public function getAge(): int
+    {
+        $birthday = $this->birthday;
+
+        if (null === $birthday) {
+            return 0;
+        }
+
+        $interval = $birthday->diff(new \DateTime());
+
+        return $interval->y;
+    }
+
+    public function getLegalRepresentative(): ?string
+    {
+        return $this->legalRepresentative;
+    }
+
+    public function setLegalRepresentative(?string $legalRepresentative): self
+    {
+        $this->legalRepresentative = $legalRepresentative;
+
+        return $this;
+    }
+
+    public function getSubscriptionDate(): ?\DateTimeInterface
+    {
+        return $this->subscriptionDate;
+    }
+
+    public function setSubscriptionDate(?\DateTimeInterface $subscriptionDate): self
+    {
+        $this->subscriptionDate = $subscriptionDate;
+
+        return $this;
+    }
+
+    public function getLaneNumber(): ?string
+    {
+        return $this->laneNumber;
+    }
+
+    public function setLaneNumber(string $laneNumber): self
+    {
+        $this->laneNumber = $laneNumber;
+
+        return $this;
+    }
+
+    public function getLaneType(): ?string
+    {
+        return $this->laneType;
+    }
+
+    public function setLaneType(string $laneType): self
+    {
+        $this->laneType = $laneType;
+
+        return $this;
+    }
+
+    public function getLaneName(): ?string
+    {
+        return $this->laneName;
+    }
+
+    public function setLaneName(string $laneName): self
+    {
+        $this->laneName = $laneName;
+
+        return $this;
+    }
+
+    public function getPostalCode(): ?int
+    {
+        return $this->postalCode;
+    }
+
+    public function setPostalCode(int $postalCode): self
+    {
+        $this->postalCode = $postalCode;
+
+        return $this;
+    }
+
+    public function getCity(): ?string
+    {
+        return $this->city;
+    }
+
+    public function setCity(string $city): self
+    {
+        $this->city = $city;
+
+        return $this;
+    }
+
+    public function getPhoneNumber(): ?string
+    {
+        return $this->phoneNumber;
+    }
+
+    public function setPhoneNumber(string $phoneNumber): self
+    {
+        $this->phoneNumber = $phoneNumber;
+
+        return $this;
+    }
+
+    public function getFormattedAddress(): string
+    {
+        $address = "{$this->getNumber()}, {$this->getLaneType()} {$this->getLaneName()}\n";
+        $address .= "{$this->getPostalCode()} {$this->getCity()}";
+
+        return $address;
+    }
+
+    /**
+     * @return Collection|LogbookEntry[]
+     */
+    public function getLogbookEntries(): Collection
+    {
+        return $this->logbookEntries;
+    }
+
+    public function addLogbookEntry(LogbookEntry $logbookEntry): self
+    {
+        if (!$this->logbookEntries->contains($logbookEntry)) {
+            $this->logbookEntries[] = $logbookEntry;
+            $logbookEntry->addCrew($this);
         }
 
         return $this;
+    }
+
+    public function removeLogbookEntry(LogbookEntry $logbookEntry): self
+    {
+        if ($this->logbookEntries->contains($logbookEntry)) {
+            $this->logbookEntries->removeElement($logbookEntry);
+            $logbookEntry->removeCrew($this);
+        }
+
+        return $this;
+    }
+
+    /**
+     * @return Collection|MedicalCertificate[]
+     */
+    public function getMedicalCertificates(): Collection
+    {
+        return $this->medicalCertificates;
+    }
+
+    public function addMedicalCertificate(MedicalCertificate $medicalCertificate): self
+    {
+        if (!$this->medicalCertificates->contains($medicalCertificate)) {
+            $this->medicalCertificates[] = $medicalCertificate;
+            $medicalCertificate->setUser($this);
+        }
+
+        return $this;
+    }
+
+    public function removeMedicalCertificate(MedicalCertificate $medicalCertificate): self
+    {
+        if ($this->medicalCertificates->contains($medicalCertificate)) {
+            $this->medicalCertificates->removeElement($medicalCertificate);
+            // set the owning side to null (unless already changed)
+            if ($medicalCertificate->getUser() === $this) {
+                $medicalCertificate->setUser(null);
+            }
+        }
+
+        return $this;
+    }
+
+    /**
+     * @Assert\Callback()
+     */
+    public function validateLegalRepresentative(ExecutionContextInterface $context, $payload)
+    {
+        if ($this->getAge() < 18) {
+            $context->buildViolation('Le membre est mineur, merci de renseigner un représentant légal.')
+                ->atPath('legalRepresentative')
+                ->addViolation();
+        }
+    }
+
+    /**
+     * @ORM\PrePersist
+     * @ORM\PreUpdate
+     */
+    public function defineUsername()
+    {
+        $slugger = new AsciiSlugger('fr');
+        $firstName = $slugger->slug($this->firstName)->lower();
+        $lastName = $slugger->slug($this->lastName)->lower();
+
+        $this->username = "$firstName.$lastName";
+    }
+
+    public static function getAvailableLicenseTypes(): array
+    {
+        return [
+            'Licence Annuelle' => self::LICENSE_TYPE_ANNUAL,
+            'Licence Indoor' => self::LICENSE_TYPE_INDOOR,
+        ];
+    }
+
+    public static function getAvailableRowerCategories(): array
+    {
+        return [
+            'A' => self::ROWER_CATEGORY_A,
+            'B' => self::ROWER_CATEGORY_B,
+            'C' => self::ROWER_CATEGORY_C,
+        ];
     }
 }
