--- src/Validator/UniqueUserMember.php
@@ -1,38 +1,0 @@
-<?php
-
-/*
- * Copyright 2020 Mathieu Piot
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-namespace App\Validator;
-
-use Symfony\Component\Validator\Constraint;
-
-/**
- * @Annotation
- */
-class UniqueUserMember extends Constraint
-{
-    /*
-     * Any public properties become valid options for the annotation.
-     * Then, use these in your validator class.
-     */
-    public $message = 'Ce membre est déjà attaché à un autre utilisateur.';
-
-    public function getTargets()
-    {
-        return self::CLASS_CONSTRAINT;
-    }
-}
