+++ templates/base.html.twig
@@ -50,7 +50,7 @@
             {% else %}
                 <div class="ml-md-auto">
                     <a href="{{ path('app_login') }}" class="btn btn-outline-secondary rounded-pill">Se connecter</a>
-                    <a href="{{ path('app_register') }}" class="btn btn-primary rounded-pill">Créer un compte</a>
+{#                    <a href="{{ path('app_register') }}" class="btn btn-primary rounded-pill">Créer un compte</a>#}
                 </div>
             {% endif %}
         </header>
