+++ templates/club_management/base.html.twig
@@ -9,9 +9,6 @@
             <a class="nav-link" href="{{ path('shell_damage_category_index') }}">Catégories d'avaries</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link" href="{{ path('member_index') }}">Membres</a>
-        </li>
-        <li class="nav-item">
             <a class="nav-link" href="{{ path('shell_index') }}">Parc</a>
         </li>
         <li class="nav-item">
