+++ templates/club_management/homepage.twig
@@ -25,15 +25,6 @@
             </div>
         </a>
 
-        <a href="{{ path('member_index') }}">
-            <div class="card h-75 m-3" style="width: 15rem;">
-                <div class="card-body text-center">
-                    <span class="fas fa-running fa-6x mb-3 text-secondary"></span>
-                    <h4 class="card-title">Membres</h4>
-                </div>
-            </div>
-        </a>
-
         <a href="{{ path('shell_index') }}">
             <div class="card h-75 m-3" style="width: 15rem;">
                 <div class="card-body text-center">
