--- templates/emails/invitation.txt.twig
@@ -1,4 +1,0 @@
-Bonjour {{ member.fullName }},
-Voici votre invitation pour vous enregistrer sur l'application web du club d'aviron de Tours.
-
-{{ invitation.code }}
