--- templates/member/_delete_form.html.twig
@@ -1,10 +1,0 @@
-<form method="post" action="{{ path('member_delete', {'id': member.id}) }}" onsubmit="return confirm('Êtes-vous sûr de vouloir supprimer cet éléménet ?');">
-    <input type="hidden" name="_method" value="DELETE">
-    <input type="hidden" name="_token" value="{{ csrf_token('delete' ~ member.id) }}">
-
-    {% block buttons %}
-        <button class="btn btn-outline-danger btn-sm">
-            <span class="fas fa-trash-alt"></span> Supprimer
-        </button>
-    {% endblock %}
-</form>
