--- templates/member/edit.html.twig
@@ -1,27 +1,0 @@
-{% extends 'club_management/base.html.twig' %}
-
-{% block title %}Modifier un membre{% endblock %}
-
-{% block breadcrumb %}
-    {{ parent() }}
-    <li class="breadcrumb-item"><a href="{{ path('member_index') }}">Gestion des membres</a></li>
-    <li class="breadcrumb-item"><a href="{{ path('member_show', {'id': member.id}) }}">{{ member.fullName }}</a></li>
-    <li class="breadcrumb-item active">Modifier un membre</li>
-{% endblock %}
-
-{% block body %}
-    <div class="d-flex">
-        <h1>Modifier un membre</h1>
-
-        <div class="ml-auto align-self-center">
-            {{ include('member/_delete_form.html.twig') }}
-        </div>
-    </div>
-
-    {{ include('member/_form.html.twig', {'button_label': 'Modifier'}) }}
-{% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-    {{ encore_entry_script_tags('member_form') }}
-{% endblock %}
