--- templates/member/index.html.twig
@@ -1,79 +1,0 @@
-{% extends 'club_management/base.html.twig' %}
-
-{% block title %}Gestion des membres{% endblock %}
-
-{% block breadcrumb %}
-    {{ parent() }}
-    <li class="breadcrumb-item active">Gestion des membres</li>
-{% endblock %}
-
-{% block body %}
-    <div class="d-flex">
-        <h1>Gestion des membres</h1>
-
-        <a href="{{ path('member_new') }}" class="btn btn-outline-primary btn-sm ml-auto align-self-center">
-            <span class="fas fa-plus"></span> Ajouter un membre
-        </a>
-    </div>
-
-    <form method="get">
-        <div class="form-row align-items-center">
-            <div class="col-md-auto col-sm-12 ml-auto">
-                <div class="input-group mb-2">
-                    <div class="input-group-prepend">
-                        <div class="input-group-text"><span class="fa fa-search"></span></div>
-                    </div>
-                    <input type="search" class="form-control" id="q" name="q" placeholder="Rechercher..." value="{{ app.request.query.alnum('q') }}">
-                </div>
-            </div>
-        </div>
-    </form>
-
-    <div class="table-responsive">
-        <table class="table table-striped">
-            <thead>
-            <tr>
-                <th>Nom</th>
-                <th>Email</th>
-                <th>Numéro de license</th>
-                <th>Licence</th>
-                <th>Date de fin de validité</th>
-                <th></th>
-            </tr>
-            </thead>
-            <tbody>
-            {% for member in members %}
-                <tr>
-                    <td>
-                        <a href="{{ path('member_show', {'id': member.id}) }}">{{ member.fullName }}</a>
-                    </td>
-                    <td>
-                        {% if member.email is not null %}
-                            <a href="mailto:{{ member.email }}">{{ member.email }}</a>
-                        {% else %}
-                            -
-                        {% endif %}
-                    </td>
-                    <td>{{ member.licenseNumber }}</td>
-                    <td>{{ member.textLicenseType }}</td>
-                    <td class="{{ member.isLicenseValid ? 'text-success' : 'text-danger' }}">{{ member.licenseEndAt ? member.licenseEndAt|date('d/m/Y') : '-' }}</td>
-                    <td class="text-right">
-                        <a class="btn btn-outline-warning btn-sm rounded-pill"
-                           href="{{ path('member_edit', {'id': member.id}) }}">
-                            <span class="fas fa-pencil-alt"></span> Modifier
-                        </a>
-                    </td>
-                </tr>
-            {% else %}
-                <tr>
-                    <td colspan="7" class="text-center">Pas de membre</td>
-                </tr>
-            {% endfor %}
-            </tbody>
-        </table>
-
-        {% if members|length < members.totalItemCount %}
-            {{ knp_pagination_render(members) }}
-        {% endif %}
-    </div>
-{% endblock %}
