--- templates/member/new.html.twig
@@ -1,20 +1,0 @@
-{% extends 'club_management/base.html.twig' %}
-
-{% block title %}Ajouter un membre{% endblock %}
-
-{% block breadcrumb %}
-    {{ parent() }}
-    <li class="breadcrumb-item"><a href="{{ path('member_index') }}">Gestion des membres</a></li>
-    <li class="breadcrumb-item active">Ajouter un membre</li>
-{% endblock %}
-
-{% block body %}
-    <h1>Ajouter un membre</h1>
-
-    {{ include('member/_form.html.twig') }}
-{% endblock %}
-
-{% block javascripts %}
-    {{ parent() }}
-    {{ encore_entry_script_tags('member_form') }}
-{% endblock %}
