--- templates/member/show.html.twig
@@ -1,150 +1,0 @@
-{% extends 'club_management/base.html.twig' %}
-
-{% block title %}{{ member.fullName }}{% endblock %}
-
-{% block breadcrumb %}
-    {{ parent() }}
-    <li class="breadcrumb-item"><a href="{{ path('member_index') }}">Gestion des membres</a></li>
-    <li class="breadcrumb-item active">{{ member.fullName }}</li>
-{% endblock %}
-
-{% block body %}
-    <div class="d-flex">
-        <h1>{{ member.fullName }}</h1>
-
-        <div class="ml-auto align-self-center ">
-            {% embed 'member/_delete_form.html.twig' %}
-                {% block buttons %}
-                    <div class="btn-group-sm">
-                        <a class="btn btn-outline-warning" href="{{ path('member_edit', {'id': member.id}) }}">
-                            <span class="fas fa-pencil-alt"></span> Modifier
-                        </a>
-
-                        {{ parent() }}
-                    </div>
-                {% endblock %}
-            {% endembed %}
-        </div>
-    </div>
-
-    <div>
-        {% if invitation is not null %}
-            <div>Invitation: {{ invitation.code }}</div>
-        {% elseif member.user is not null %}
-            <div>Utilisateur: {{ member.user.email }}</div>
-        {% else %}
-            <a class="btn btn-sm btn-outline-primary" href="{{ path('invitation_new', { 'id': member.id }) }}">
-                <span class="fa fa-paper-plane"></span> Envoyer une invitation
-            </a>
-        {% endif %}
-        <div class="mb-3">Date d'inscription: {{ member.subscriptionDate|date('d/m/Y') }}</div>
-
-        <div class="row">
-            <div class="col-xl">
-                <div class="card">
-                    <div class="card-header">
-                        Membre
-                    </div>
-                    <ul class="list-group list-group-flush">
-                        <li class="list-group-item d-flex justify-content-between">
-                            Genre
-                            <span>{{ member.textGender }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Nom
-                            <span>{{ member.fullName }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Email
-                            <span>{{ member.email }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Date de naissance
-                            <span>{{ member.birthday|date('d/m/Y') }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Représentant légal
-                            <span>{{ member.legalRepresentative ? member.legalRepresentative : '-' }}</span>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-
-            {% if member.address is not null %}
-                <div class="col-xl mt-3 mt-xl-0">
-                    <div class="card">
-                        <div class="card-header">
-                            Adresse
-                        </div>
-                        <ul class="list-group list-group-flush">
-                            <li class="list-group-item d-flex justify-content-between">
-                                Adresse
-                                <span>{{ member.address.formattedAddress|nl2br }}</span>
-                            </li>
-                            <li class="list-group-item d-flex justify-content-between">
-                                Numéro de téléphone
-                                <span>{{ member.address.phoneNumber }}</span>
-                            </li>
-                        </ul>
-                    </div>
-                </div>
-            {% endif %}
-        </div>
-
-        <div class="row mt-3">
-            <div class="col-xl">
-                <div class="card">
-                    <div class="card-header">
-                        Licence
-                    </div>
-                    <ul class="list-group list-group-flush">
-                        <li class="list-group-item d-flex justify-content-between">
-                            Type de licence
-                            <span>{{ member.textLicenseType }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Date de fin de validité
-                            <span>{{ member.licenseEndAt ? member.licenseEndAt|date('d/m/Y') : '-' }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Numéro de licence
-                            <span>{{ member.licenseNumber ? member.licenseNumber : '-' }}</span>
-                        </li>
-                        <li class="list-group-item d-flex justify-content-between">
-                            Catégorie rameur
-                            <span>{{ member.textRowerCategory }}</span>
-                        </li>
-                    </ul>
-                </div>
-            </div>
-
-            <div class="col-xl mt-3 mt-xl-0">
-                <div class="card">
-                    <div class="card-header">
-                        Certificats médicaux
-                    </div>
-                    <div class="card-body p-0">
-                        <table class="table table-striped">
-                            <thead>
-                            <tr>
-                                <th>Date</th>
-                                <th>Type</th>
-                                <th>Niveau</th>
-                            </tr>
-                            </thead>
-                            <tbody>
-                            {% for medicalCertificate in member.medicalCertificates %}
-                                <tr>
-                                    <td>{{ medicalCertificate.date|date('d/m/Y') }}</td>
-                                    <td>{{ medicalCertificate.textType }}</td>
-                                    <td>{{ medicalCertificate.textLevel }}</td>
-                                </tr>
-                            {% endfor %}
-                            </tbody>
-                        </table>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-{% endblock %}
