+++ templates/security/login.html.twig
@@ -22,8 +22,8 @@
         {% endif %}
 
         <div class="form-label-group">
-            <input type="email" value="{{ last_username }}" name="email" id="inputEmail" class="form-control" placeholder="Email" required autofocus>
-            <label for="inputUsername">Email</label>
+            <input type="text" value="{{ last_username }}" name="username" id="inputEmail" class="form-control" placeholder="Utilisateur" required autofocus>
+            <label for="inputUsername">Utilisateur</label>
         </div>
 
         <div class="form-label-group">
