--- templates/security/register.html.twig
@@ -1,33 +1,0 @@
-{% extends 'security/_base.html.twig' %}
-{% form_theme registrationForm _self %}
-
-{% block title %}Créer un compte{% endblock %}
-
-{% block form_row -%}
-    {%- set widget_attr = {} -%}
-    <{{ element|default('div') }}{% with {attr: row_attr|merge({class: (row_attr.class|default('') ~ ' form-label-group')|trim})} %}{{ block('attributes') }}{% endwith %}>
-        {{- form_widget(form, widget_attr) -}}
-        {{- form_label(form) -}}
-        {{- form_errors(form) -}}
-        {{- form_help(form) -}}
-    </{{ element|default('div') }}>
-{%- endblock form_row %}
-
-{% block body_content %}
-    {{ form_start(registrationForm) }}
-        {{ form_row(registrationForm.email, {
-            'attr': { 'placeholder': 'Email' }
-        }) }}
-        {{ form_row(registrationForm.plainPassword.first, {
-            'attr': { 'placeholder': 'Mot de passe' }
-        }) }}
-        {{ form_row(registrationForm.plainPassword.second, {
-            'attr': { 'placeholder': 'Répéter le mot de passe' }
-        }) }}
-        {{ form_row(registrationForm.invitation, {
-            'attr': { 'placeholder': 'Invitation' }
-        }) }}
-
-    <button class="btn btn-lg btn-primary btn-block" type="submit">Créer mon compte</button>
-    {{ form_end(registrationForm) }}
-{% endblock %}
