+++ templates/user/_form.html.twig
@@ -1,11 +1,96 @@
 {{ form_start(form) }}
-    {{ form_row(form.email) }}
-    {% if form.plainPassword is defined%}
-        {{ form_row(form.plainPassword) }}
-    {% endif %}
-    {% if form.roles is defined%}
-        {{ form_row(form.roles) }}
-    {% endif %}
+    {{ form_row(form.subscriptionDate) }}
 
-    <button class="btn btn-primary">{{ button_label|default('Sauver') }}</button>
+<div class="row">
+    <div class="col-xl">
+        <div class="card">
+            <div class="card-header">
+                Membre
+            </div>
+            <div class="card-body">
+                {{ form_row(form.gender) }}
+                <div class="row">
+                    <div class="col">
+                        {{ form_row(form.firstName) }}
+                    </div>
+                    <div class="col">
+                        {{ form_row(form.lastName) }}
+                    </div>
+                </div>
+                {{ form_row(form.email) }}
+                {% if form.plainPassword is defined%}
+                    {{ form_row(form.plainPassword) }}
+                {% endif %}
+                {% if form.roles is defined%}
+                    {{ form_row(form.roles) }}
+                {% endif %}
+                {{ form_row(form.birthday) }}
+                {{ form_row(form.legalRepresentative) }}
+            </div>
+        </div>
+    </div>
+
+    <div class="col-xl mt-3 mt-xl-0">
+        <div class="card">
+            <div class="card-header">
+                Adresse
+            </div>
+            <div class="card-body">
+                <div class="row">
+                    <div class="col">
+                        {{ form_row(form.address.laneNumber) }}
+                    </div>
+                    <div class="col">
+                        {{ form_row(form.address.laneType) }}
+                    </div>
+                    <div class="col">
+                        {{ form_row(form.address.laneName) }}
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col">
+                        {{ form_row(form.address.postalCode) }}
+                    </div>
+                    <div class="col">
+                        {{ form_row(form.address.city) }}
+                    </div>
+                </div>
+
+                {{ form_row(form.address.phoneNumber) }}
+            </div>
+        </div>
+    </div>
+</div>
+
+<div class="card mt-3">
+    <div class="card-header">
+        Licence
+    </div>
+    <div class="card-body">
+        <div class="row">
+            <div class="col">
+                {{ form_row(form.licenseType) }}
+            </div>
+            <div class="col">
+                {{ form_row(form.licenseEndAt) }}
+            </div>
+            <div class="col">
+                {{ form_row(form.licenseNumber) }}
+            </div>
+        </div>
+
+        {{ form_row(form.rowerCategory) }}
+    </div>
+</div>
+
+<div class="card mt-3">
+    <div class="card-header">
+        Certificats
+    </div>
+    <div class="card-body">
+        {{ form_row(form.medicalCertificates) }}
+    </div>
+</div>
+
+<button class="btn btn-primary mt-3">{{ button_label|default('Sauver') }}</button>
 {{ form_end(form) }}
