+++ webpack.config.js
@@ -38,7 +38,6 @@ Encore
      */
     .addEntry('app', './assets/js/app.js')
     .addEntry('logbook_entry_form', './assets/js/logbook_entry_form.js')
-    .addEntry('member_form', './assets/js/member_form.js')
     .addEntry('user_form', './assets/js/user_form.js')
     //.addEntry('page2', './assets/js/page2.js')
